Safari-611.1.10/LayoutTests/fast/forms/watchos/username-text-content-type-expected.txt

PASS textContentTypeForLoneTextField is ''
PASS textContentTypeForTextFieldBeforePassword is 'username'
PASS textContentTypeForTextFieldWithAutocompleteUsername is 'username'
PASS successfullyParsed is true

TEST COMPLETE
